DROID-3484 fix
